complete project restructuring
