+++ spam_classification/graphs/KNN_learning_curves_k_1.png
