+++ spam_classification/graphs/NB_learning_curve_alpha_4.png
