+++ spam_classification/graphs/NB_learning_curve_old.png
